safeguard lxc calls

+++ .chezmoiscripts/run_9_cleanup_lxc.sh
@@ -1,4 +1,6 @@
-#!/bin/sh
-lxc project switch snapcraft
+#!/bin/bash
 
-lxc list -f compact | awk '{print $1;}' | tail -n+2 | xargs -r lxc delete
+if [[ -n "$(lxc project list -f compact | grep snapcraft)" ]]; then
+    lxc project switch snapcraft
+    lxc list -f compact | awk '{print $1;}' | tail -n+2 | xargs -r lxc delete
+fi
